Fix critical production crashes and robustify Firestore sync. - Hardened data models with @Keep and default values for Firestore compatibility. - Fixed navigation redirection race conditions in WeatherPremiumActivity. - Robustified Firestore listeners in ThemeViewModel and TravelViewModel. - Updated ProGuard rules to prevent stripping of data classes. - Fixed type mismatches and Composable lifecycle issues in detail screens.

+++ mobile/android/app/src/main/java/com/havamania/LocationTracker.kt
@@ -74,7 +74,7 @@ class DefaultLocationTracker(
                     val districtName = address.locality ?: address.subLocality ?: address.subAdminArea
 
                     GeocodingResultDto(
-                        id = (location.latitude * 1000 + location.longitude * 1000).toInt(),
+                        id = (location.latitude * 1000 + location.longitude * 1000).toLong(),
                         name = districtName ?: cityName,
                         latitude = location.latitude,
                         longitude = location.longitude,
